#45 rename package sudoku.invoker to sudoku.command
